chore: move revoke phone number to own component
